Improved prepare datasets process (with CLI)
