+++ train.py
@@ -5,7 +5,7 @@ from engine import Trainer
 
 
 if __name__ == '__main__':
-    parser = argparse.ArgumentParser(description='Script options')
+    parser = argparse.ArgumentParser(description='Train Script Options')
     parser.add_argument('--pretrain_checkpoint', type=str, default=None, help='Pretrain checkpoint to load.')
     parser.add_argument('--instruct_checkpoint', type=str, default=None, help='Instruct checkpoint to load.')
     parser.add_argument('--dpo_checkpoint', type=str, default=None, help='DPO checkpoint to load.')
